feat: extract *git dirty* check
